+++ components/xgimi_remote/__init__.py
@@ -10,6 +10,7 @@ CONF_BLE_SERVER_ID = "ble_server_id"
 CONF_KEYBOARD_REPORT_ID = "keyboard_report_id"
 CONF_CONSUMER_REPORT_ID = "consumer_report_id"
 CONF_REMOTE_NAME = "remote_name"
+CONF_TAP_DURATION_MS = "tap_duration_ms"
 CONF_WAKE_TOKEN = "wake_token"
 
 xgimi_remote_ns = cg.esphome_ns.namespace("xgimi_remote")
@@ -29,6 +30,7 @@ CONFIG_SCHEMA = cv.Schema(
         cv.Optional(CONF_REMOTE_NAME, default="M5Stack Atom Lite"): cv.All(
             cv.string_strict, cv.Length(min=1, max=20)
         ),
+        cv.Optional(CONF_TAP_DURATION_MS, default=100): cv.int_range(min=20, max=1000),
         cv.Required(CONF_WAKE_TOKEN): cv.All(
             cv.ensure_list(cv.hex_uint8_t), cv.Length(min=15, max=15)
         ),
@@ -50,6 +52,7 @@ async def to_code(config):
     cg.add(var.set_keyboard_report(keyboard_report))
     cg.add(var.set_consumer_report(consumer_report))
     cg.add(var.set_remote_name(config[CONF_REMOTE_NAME]))
+    cg.add(var.set_tap_duration_ms(config[CONF_TAP_DURATION_MS]))
     cg.add(var.set_wake_token(config[CONF_WAKE_TOKEN]))
 
     esp32_ble.register_gap_event_handler(ble, var)